CRDL-250 Remove the API-related stuff, set up scalafmt, update dependencies
