+++ app/uk/gov/hmrc/crdlcache/controllers/CodeListsController.scala
@@ -14,14 +14,17 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.crdlcache.config
+package uk.gov.hmrc.crdlcache.controllers
 
-import com.google.inject.AbstractModule
+import uk.gov.hmrc.play.bootstrap.backend.controller.BackendController
+import play.api.mvc.{Action, AnyContent, ControllerComponents}
+import javax.inject.{Inject, Singleton}
+import scala.concurrent.Future
 
-class Module extends AbstractModule {
+@Singleton()
+class CodeListsController @Inject() (cc: ControllerComponents) extends BackendController(cc) {
 
-  override def configure(): Unit = {
-
-    bind(classOf[AppConfig]).asEagerSingleton()
+  def hello(): Action[AnyContent] = Action.async { implicit request =>
+    Future.successful(Ok("Hello world"))
   }
 }
