+++ conf/app.routes
@@ -1,3 +1,3 @@
 # microservice specific routes
 
-GET        /hello-world             uk.gov.hmrc.crdlcache.controllers.MicroserviceHelloWorldController.hello
+GET        /hello-world             uk.gov.hmrc.crdlcache.controllers.CodeListsController.hello()
