--- conf/definition.routes
@@ -1,2 +1,0 @@
-GET        /api/definition                 uk.gov.hmrc.crdlcache.controllers.DocumentationController.definition()
-GET        /api/conf/:version/*file        uk.gov.hmrc.crdlcache.controllers.DocumentationController.specification(version, file)
