+++ conf/prod.routes
@@ -1,4 +1,3 @@
 # Add all the application routes to the app.routes file
 ->         /                          app.Routes
 ->         /                          health.Routes
-->         /                          definition.Routes
