--- resources/public/api/conf/common/overview.md
@@ -1,1 +1,0 @@
-This is a placeholder overview.md for a new API Microservice repository
