--- resources/public/api/definition.json
@@ -1,17 +1,0 @@
-{
-  "scopes": [
-  ],
-  "api": {
-    "name": "Hello World",
-    "description": "A 'hello world' example of an API on the HMRC API Developer Hub.",
-    "context": "discuss-with-the-api-platform-team",
-    "categories": ["OTHER"],
-    "versions": [
-      {
-        "version": "1.0",
-        "status": "ALPHA",
-        "endpointsEnabled": false
-      }
-    ]
-  }
-}
